Refactor

+++ CodeTest/Rates/IUniRateService.cs
@@ -4,6 +4,6 @@ namespace CodeTest.Rates
 {
     public interface IUniRateService
     {
-        Task<decimal> GetRate(Currency buyCurrency, Currency sellCurrency);
+        Task<double> GetRate(Currency buyCurrency, Currency sellCurrency);
     }
 }
